+++ Dfe.Academies.Academisation.Domain.UnitTest/ConversionAdvisoryBoardDecisionAggregate/ConversionAdvisoryBoardDecisionSetIdTests.cs
@@ -15,8 +15,8 @@ public class ConversionAdvisoryBoardDecisionSetIdTests
 	{
 		//Arrange
 		var details = _fixture.Create<AdvisoryBoardDecisionDetails>();
-		ConversionAdvisoryBoardDecision target = new(1, details, null, null, null, default, default);
-		
+		ConversionAdvisoryBoardDecision target = new(1, details, null, null, null, null, default, default);
+
 		//Act & Assert
 		Assert.Throws<InvalidOperationException>(() => target.SetId(1));
 	}
@@ -31,8 +31,9 @@ public class ConversionAdvisoryBoardDecisionSetIdTests
 		var deferred = _fixture.CreateMany<AdvisoryBoardDeferredReasonDetails>();
 		var declined = _fixture.CreateMany<AdvisoryBoardDeclinedReasonDetails>();
 		var withdrawn = _fixture.CreateMany<AdvisoryBoardWithdrawnReasonDetails>();
-		ConversionAdvisoryBoardDecision expected = new(expectedId, details, deferred, declined, withdrawn, default, default);
-		ConversionAdvisoryBoardDecision target = new(default, details, deferred, declined, withdrawn, default, default);
+		var daoRevokedReasons = _fixture.CreateMany<AdvisoryBoardDAORevokedReasonDetails>();
+		ConversionAdvisoryBoardDecision expected = new(expectedId, details, deferred, declined, withdrawn, daoRevokedReasons, default, default);
+		ConversionAdvisoryBoardDecision target = new(default, details, deferred, declined, withdrawn, daoRevokedReasons, default, default);
 
 		//Act
 		target.SetId(expectedId);