+++ Dfe.Academies.Academisation.Service.UnitTest/Commands/TransferProject/SetTransferProjectDeleteCommandHandlerTests.cs
@@ -1,0 +1,93 @@
+﻿using Dfe.Academies.Academisation.Core;
+using Dfe.Academies.Academisation.Domain.SeedWork;
+using Dfe.Academies.Academisation.Domain.TransferProjectAggregate;
+using Dfe.Academies.Academisation.Service.Commands.Application;
+using Dfe.Academies.Academisation.Service.Commands.TransferProject;
+using FluentAssertions;
+using Microsoft.Extensions.Logging;
+using Moq;
+using Xunit;
+
+namespace Dfe.Academies.Academisation.Service.UnitTest.Commands.TransferProject
+{
+	public class SetTransferProjectDeleteCommandHandlerTests
+	{
+		private readonly Mock<ITransferProjectRepository> _transferProjectRepositoryMock;
+		private readonly Mock<ILogger<SetTransferProjectDeletedAtCommandHandler>> _loggerMock;
+		private readonly SetTransferProjectDeletedAtCommandHandler _handler;
+
+		public SetTransferProjectDeleteCommandHandlerTests()
+		{
+			_transferProjectRepositoryMock = new Mock<ITransferProjectRepository>();
+			_loggerMock = new Mock<ILogger<SetTransferProjectDeletedAtCommandHandler>>();
+			_handler = new SetTransferProjectDeletedAtCommandHandler(_transferProjectRepositoryMock.Object, _loggerMock.Object);
+		}
+
+		[Fact]
+		public async Task Handle_TransferProjectNotFound_ReturnsNotFoundCommandResult()
+		{
+			// Arrange
+			var command = new SetTransferProjectDeletedAtCommand(1);
+
+			_transferProjectRepositoryMock.Setup(x => x.GetById(It.IsAny<int>()))!
+				.ReturnsAsync((Domain.TransferProjectAggregate.TransferProject)null);
+
+			// Act
+			var result = await _handler.Handle(command, default);
+
+			// Assert
+			result.Should().BeOfType<NotFoundCommandResult>();
+			_loggerMock.Verify(logger =>
+					logger.Log(
+						LogLevel.Error,
+						It.IsAny<EventId>(),
+						It.Is<It.IsAnyType>((v, t) => true),
+						It.IsAny<Exception>(),
+						It.Is<Func<It.IsAnyType, Exception, string>>((v, t) => true)!),
+				Times.Once);
+		}
+
+		[Fact]
+		public async Task Handle_TransferProjectFound_ReturnsCommandSuccessResult()
+		{
+			// Arrange
+			var command = new SetTransferProjectDeletedAtCommand(1);
+
+			// Arrange      
+			string outgoingTrustUkprn = "11112222";
+			string outgoingTrustName = "outgoingTrustName";
+			string incomingTrustUkprn = "11110000";
+			string incomingTrustName = "incomingTrustName";
+			List<TransferringAcademy> academies = new List<TransferringAcademy>() { new TransferringAcademy(incomingTrustUkprn, incomingTrustName, "22221111", "region", "local authority") };
+
+			bool isFormAMat = true;
+			DateTime createdOn = DateTime.Now;
+
+			// Create a transfer project
+			var transferProject = Domain.TransferProjectAggregate.TransferProject.Create(outgoingTrustUkprn,
+				outgoingTrustName,
+				academies,
+				isFormAMat,
+				createdOn);
+			// Mock Unit of work and Repository 
+			var unitOfWorkMock = new Mock<IUnitOfWork>();
+			unitOfWorkMock.Setup(u => u.SaveChangesAsync(It.IsAny<CancellationToken>()))
+				.Returns(Task.FromResult(1));  // Return Task<int> with a dummy value
+			_transferProjectRepositoryMock.Setup(repo => repo.UnitOfWork)
+				.Returns(unitOfWorkMock.Object);
+
+			// Mock GetById to use our Transfer Project from above
+			_transferProjectRepositoryMock.Setup(x => x.GetByUrn(It.IsAny<int>())).ReturnsAsync(transferProject);
+
+			// Act
+			var result = await _handler.Handle(command, new CancellationToken(false));
+
+			// Assert
+			result.Should().BeOfType<CommandSuccessResult>();
+
+
+			_transferProjectRepositoryMock.Verify(repo => repo.Update(It.IsAny<Domain.TransferProjectAggregate.TransferProject>()), Times.Once);
+		}
+
+	}
+}