+++ release-notes.md
@@ -1,6 +1,7 @@
 ## NEXT
 * Added ability to edit PFI schemes within a Transfer project for their requisite Transferring Academies
 * Added Region and Local authority to the transfering academy record for filtering purposes
+* Added ability to have 'DAO Revoked' decision recorded against a Sponsored conversion project
 ## 9.2.0
 * Added transfer projects
 * Added fields required to a Conversion Project to allow for sponsored grants